Move serial implementation code out of macro
Having all this code in a macro has numerous disadvantages:
- The code is less readable, as you often have to jump between the code
  itself and the arguments passed into the macro.
- The code is deeply indented, and the macro can sometime confuse the
  syntax highlighting, both of which exacerbate the readability problem.
- Any compiler error in the macro shows up multiple times, for each
  implementation that the macro generates.
- Multiple nearly identical methods are generated. I haven't checked
  whether the affects the size of the compiled binary, but even if the
  compiler is smart enough to optimize this, why not unify the common
  code into one method.

I'm sure there are more, but this is what I can think of off the top of
my head.

This commit replaces the big-macro approach with the following:
- The bulk of the code is moved out of the macro.
- The out-of-macro code is made generic over the USART instance.
- A trait is implemented for all USART instances, which abstracts over
  the differences between them.
- The implementations of this traits are the last thing left in the
  macro.

I realize this breaks with tradition, but for the reasons mentioned, I
think this approach is superior to the classic "everything in the macro"
approach. There's also prior art in the nrf-rs[1] and lpc-rs[2] worlds.

[1] https://github.com/nrf-rs/
[2] https://github.com/lpc-rs/