Make DMA API more flexible, less safe
To support typical SPI use cases, it must be possible to have two
ongoing DMA transfers that both read from/write to the same buffer. This
is not possible with the current DMA abstractions.

This commit solves that problem, by relaxing the requirements on the
creation of a new `Transfer`. As a consequence, the safe constructors of
`Transfer` have been removed, and only the unsafe variant is available
now. This is unfortunate, but I believe it is necessary.

Please note that the burden of calling the unsafe constructor is put on
the peripheral APIs. This doesn't change anything for the end user, who
will still use the same (safe) APIs to start DMA transfers.