Add dedicated `Enabled` state struct for SPI
The difference to `state::Enabled` is that `spi::Enabled` tracks the
word size the SPI instance was configured for. This is required to add
support for other word sizes. Specifically, I'm working on adding
support for 16-bit writes.

This commit doesn't add support for other word sized yet. It just
replaces `state::Enabled` with `spi::Enabled<u8>` where applicable.