Fix 16-bit support for SPI
There was a configuration error in the SPI module that caused 16-bit
mode to not work correctly. The 16-bit words were basically treated as
two bytes. This was indicated by the fact that I had to convert the word
in the example to big endian to get the right result.

When I first discovered that this conversion was necessary, I was
suprised, but eventually convinced myself that this was normal. Well, it
isn't, and the way it works now makes a lot more sense.