Implement full I2C timing using Arduino algorithm
The Arduino project on has a very nice algorithm for determining STM32
I2C timing parameters based on the I2C timing specs for each mode. I
ported the algorithm to Rust, and Rust-ified it a bit. The C origins
show through, but it works well. Test for Standard and Fast modes on
real hardware.