Initial Github Actions CI
This patch provides an initial port of the current Travis CI
configuration to Github Actions to improve our CI overall and be more
consistent with the other stm32-hal crates. It also enables a basic
`clippy` lint check on the crate. At the moment, there are quite a few
warnings and errors that should be fixed in the future. Travis CI
support has been left unmodified for now, but should be removed later
too.