Fix mass-erase trigger on "small" devices.
The list of specifically enumerated "large" devices has the correct set
of register operations to start a mass erase, setting the mer1 and mer2
bits.  But the "small-devices" alternative does not; it clears the mer
bit instead of setting it.

Looking at history, this was changed from set to clear in the PR that
added support for "large" devices.  It appears to have been accidental.