Fix ref-to-mut-static warning in serial_dma
This is mostly a workaround, because Pin<> requires a reference, not a
pointer, but the rust compiler change wants users to use pointers (via
addr_of_mut!()), not references.  In this particular case, we don't
actually *need* Pin<>, because the static mut won't be moved anyway, but
we can't prove that to the DMA API without it.

So deref the addr_of_mut!() pointer, and make a ref to it, rather than
restructuring the DMA API to work on raw pointers (and somehow enforce
the Pin<> guarantees on the pointed-to data).