Use fixed-point u64 math to divide the PLL values
This avoids the need for almost 1K of code to implement division for u64
values (which the Cortex-M7 can't do in hardware).  It can multiply u64
values in hardware, so instead of dividing by (integer) x, we find 1/x
in fixed-point, multiply, and shift the fixed-point offset back out.
Finding 1/x can be done in a u32 since our shift is small enough, and
that division *can* be done in hardware on this CPU.

Fixes #211