Round 1/x, and use 30-bit fixed point for them.
This requires shifting the input frequencies down by 4 bits in order to
fit both the max base_clk*n value, and the 30-bit fractional part, into
a u64.  (base_clk*n is up to 38 bits wide.)  This is OK as long as all
input frequencies are a multiple of 250kHz, which has the bottom 4 bits
clear.